[Fix] Arrow Run-End Encoded INT64 run_ends use wrong template parameter
The INT64 case in ColumnArrowToDuckDBRunEndEncoded incorrectly
dispatched to FlattenRunEndsSwitch<int32_t> instead of
FlattenRunEndsSwitch<int64_t>, a copy-paste error from the INT32 case.
This caused corrupted data when scanning Arrow REE columns with int64
run_ends, as the int64 buffer was reinterpreted with int32 stride.

+++ src/function/table/arrow_conversion.cpp
@@ -736,7 +736,7 @@ void ArrowToDuckDBConversion::ColumnArrowToDuckDBRunEndEncoded(Vector &vector, c
 		FlattenRunEndsSwitch<int32_t>(vector, run_end_encoding, compressed_size, scan_offset, size);
 		break;
 	case PhysicalType::INT64:
-		FlattenRunEndsSwitch<int32_t>(vector, run_end_encoding, compressed_size, scan_offset, size);
+		FlattenRunEndsSwitch<int64_t>(vector, run_end_encoding, compressed_size, scan_offset, size);
 		break;
 	default:
 		throw NotImplementedException("Type '%s' not implemented for RunEndEncoding", TypeIdToString(physical_type));
